pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "leetcode-auto"
version = "1.0.0"
description = "LeetCode Hot100 每日同步工具 - 自动获取刷题记录并更新每日计划"
requires-python = ">=3.9"
dependencies = [
    "requests",
    "python-dotenv",
    "playwright",
    "schedule",
    "rich",
]

[project.scripts]
leetcode = "leetcode_auto.sync:main"
